requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.3
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
googleapis-common-protos==1.70.0
gotrue==2.12.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.8.2
kombu==5.4.2
Markdown==3.7
multidict==6.4.4
nanoid==2.0.0
openai==1.64.0
packaging==25.0
pluggy==1.6.0
postgrest==1.0.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.3
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
yarl==1.20.0
